chore: Refactor move api ITransactionDetailsRequest/Response into chrysalis folder and fix imports
