+++ api/src/models/api/chrysalis/ITransactionsDetailsRequest.ts
