+++ api/src/models/api/chrysalis/ITransactionsDetailsResponse.ts
@@ -1,4 +1,4 @@
-import { IResponse } from "./IResponse";
+import { IResponse } from "../IResponse";
 
 export interface Output {
     output?: {
