+++ api/src/routes/chrysalis/transactionhistory/get.ts
@@ -1,6 +1,6 @@
 import { ServiceFactory } from "../../../factories/serviceFactory";
-import { ITransactionsDetailsRequest } from "../../../models/api/ITransactionsDetailsRequest";
-import { ITransactionsDetailsResponse } from "../../../models/api/ITransactionsDetailsResponse";
+import { ITransactionsDetailsRequest } from "../../../models/api/chrysalis/ITransactionsDetailsRequest";
+import { ITransactionsDetailsResponse } from "../../../models/api/chrysalis/ITransactionsDetailsResponse";
 import { IConfiguration } from "../../../models/configuration/IConfiguration";
 import { CHRYSALIS } from "../../../models/db/protocolVersion";
 import { NetworkService } from "../../../services/networkService";
