+++ api/src/utils/chrysalis/chrysalisTangleHelper.ts
@@ -6,8 +6,8 @@ import { ChronicleClient } from "../../clients/chronicleClient";
 import { HornetClient } from "../../clients/hornetClient";
 import { IMessageDetailsResponse } from "../../models/api/chrysalis/IMessageDetailsResponse";
 import { ISearchResponse } from "../../models/api/chrysalis/ISearchResponse";
-import { ITransactionsDetailsRequest } from "../../models/api/ITransactionsDetailsRequest";
-import { ITransactionsDetailsResponse } from "../../models/api/ITransactionsDetailsResponse";
+import { ITransactionsDetailsRequest } from "../../models/api/chrysalis/ITransactionsDetailsRequest";
+import { ITransactionsDetailsResponse } from "../../models/api/chrysalis/ITransactionsDetailsResponse";
 import { ITransactionsCursor } from "../../models/api/og/ITransactionsCursor";
 import { TransactionsGetMode } from "../../models/api/og/transactionsGetMode";
 import { INetwork } from "../../models/db/INetwork";
